+++ pubspec.yaml
@@ -51,6 +51,7 @@ dependencies:
   flutter_spinkit: ^5.1.0
   auto_size_text: ^3.0.0
   boxicons: ^2.1.3
+  flutter_secure_storage: ^8.0.0
 
 dev_dependencies:
   flutter_test:
